+++ tests/unit/test_xdocker.py
@@ -1,4 +1,4 @@
-"""Test."""
+"""Unit tests for the xdocker module."""
 
 from subprocess import CalledProcessError
 from unittest.mock import Mock, patch
@@ -10,7 +10,8 @@ from hamcrest import (
 )
 
 from pytest_xdocker.docker import DockerContainer
-from pytest_xdocker.docker_xrun import (
+from pytest_xdocker.xdocker import (
+    docker_call,
     docker_remove,
     docker_run,
     main,
@@ -24,42 +25,38 @@ def test_docker_remove_non_existing(unique):
     docker_remove(name)
 
 
-@patch("pytest_xdocker.docker_xrun.docker")
-def test_docker_run_container(docker_mock, unique):
+def test_docker_call(unique):
+    """Calling a command other than run should raise."""
+    with pytest.raises(ValueError):
+        docker_call("rm")
+
+
+def test_docker_run_container(unique):
     """Running a docker container should return its name."""
     name = unique("text")
-    docker_mock.command = Mock(
-        return_value=Mock(
-            with_optionals=Mock(
-                return_value=Mock(with_positionals=Mock(return_value=Mock(execute=Mock(return_value=f"{name}\n"))))
-            )
+    command = Mock(
+        with_optionals=Mock(
+            return_value=Mock(with_positionals=Mock(return_value=Mock(execute=Mock(return_value=f"{name}\n"))))
         )
     )
-    result = docker_run()
+    result = docker_run(command=command)
     assert result == name
 
 
-@patch("pytest_xdocker.docker_xrun.docker")
-def test_docker_run_already_in_use(docker_mock, unique):
+def test_docker_run_already_in_use(unique):
     """Running a docker container already in use should return None."""
-    unique("text")
-    docker_mock.command = Mock(
-        side_effect=CalledProcessError(1, "", output='The container name "/{name}" is already in use')
+    name = unique("text")
+    command = Mock(
+        with_optionals=Mock(
+            side_effect=CalledProcessError(1, "", output=f'The container name "/{name}" is already in use')
+        )
     )
-    result = docker_run()
+    result = docker_run(command=command)
     assert result is None
 
 
-@patch("pytest_xdocker.docker_xrun.docker")
-def test_docker_run_error(docker_mock, unique):
-    """Encountering an error when running a docker container should raise."""
-    docker_mock.command = Mock(side_effect=CalledProcessError(1, "", output=""))
-    with pytest.raises(CalledProcessError):
-        docker_run()
-
-
-@patch("pytest_xdocker.docker_xrun.DockerContainer")
-@patch("pytest_xdocker.docker_xrun.check_call")
+@patch("pytest_xdocker.xdocker.DockerContainer")
+@patch("pytest_xdocker.xdocker.check_call")
 def test_monitor_container_normal_exit(cc_mock, dc_mock, unique):
     """When the check_call exits, if container has no status exit.
 
@@ -77,8 +74,8 @@ def test_monitor_container_normal_exit(cc_mock, dc_mock, unique):
     dc_mock.assert_called_once_with(name)
 
 
-@patch("pytest_xdocker.docker_xrun.DockerContainer")
-@patch("pytest_xdocker.docker_xrun.check_call")
+@patch("pytest_xdocker.xdocker.DockerContainer")
+@patch("pytest_xdocker.xdocker.check_call")
 def test_monitor_container_follow_error(cc_mock, dc_mock, unique):
     """If the --follow command fails but container is still up, continue."""
     name = unique("text")
@@ -99,7 +96,7 @@ def test_monitor_container_follow_error(cc_mock, dc_mock, unique):
     assert_that(dc_mock, has_properties(call_count=2))
 
 
-@patch("pytest_xdocker.docker_xrun.check_call")
+@patch("pytest_xdocker.xdocker.check_call")
 def test_monitor_container_any_error(cc_mock, unique):
     """If the --follow command fails but container is still up, continue."""
     cc_mock.side_effect = ValueError
@@ -108,9 +105,10 @@ def test_monitor_container_any_error(cc_mock, unique):
         monitor_container(unique("text"), interval=0)
 
 
-def test_main_detach():
+def test_main_detach(capsys):
     """The main function should output an error when trying to detach."""
-    with pytest.raises(SystemExit) as e:
-        main(["--detach"])
+    with pytest.raises(SystemExit):
+        main(["run", "--detach"])
 
-    assert "Cannot pass --detach" in str(e.value)
+    captured = capsys.readouterr()
+    assert "Cannot pass --detach" in captured.err